Basic scaffold for VegRev App
This is adding the basic structure for future expansion of the app. Routes are loaded with prefixes (each prefix in it's own file) All Models are in the one file - if I end up with so many that it's unmanageable I will move to DBIC. I've done it this way to encourage myself to move to a proper ORM, instead of continuing to hack on my hack when it gets big.